refactor: split canvas-toolbar into individual tool entries
canvas-toolbar is no longer a monolithic composite. Each canvas tool
is now its own entry in toolbar.config.json on the canvas-toolbar
surface, with its own Svelte component:

- canvas-zoom: CanvasZoomControl.svelte
- canvas-zoom-to-fit: CanvasZoomToFit.svelte
- canvas-undo-redo: CanvasUndoRedo.svelte
- canvas-snap: CanvasSnap.svelte

All share the canvas-toolbar handler (actions + event bridge).
CoreUIBar passes render type to component() for dynamic resolution.
Deleted monolithic CanvasToolbar.svelte.
