+++ src/canvas/button-patterns.canvas.jsonl
@@ -301,3 +301,5 @@
 {"event":"widgets_replaced","timestamp":"2026-04-10T14:03:28.545Z","widgets":[{"id":"sticky-1","type":"sticky-note","position":{"x":283,"y":143},"props":{"text":"Primary buttons should only appear once per page section","color":"blue","width":512,"height":64}},{"id":"md-1","type":"markdown","position":{"x":127,"y":911},"props":{"content":"## Usage Guidelines\n\n- Use **primary** for the main action!!!\n- Use **danger** only for destructive actions\n- Always pair danger with a confirmation dialog","width":400}},{"id":"sticky-note-oxe2rj","type":"sticky-note","position":{"x":1701,"y":92},"props":{"text":"Ensure docs are up to date","color":"orange","width":333,"height":157}},{"id":"prototype-wltmlu","type":"prototype","position":{"x":1111,"y":605},"props":{"src":"/Example?flow=Example%2Fdefault","label":"","zoom":100,"width":800,"height":600}}]}
 {"event":"source_updated","timestamp":"2026-04-10T14:03:31.504Z","sources":[{"export":"PrimaryButtons","position":{"x":1046,"y":1831}},{"export":"DangerButtons","position":{"x":417,"y":1058},"width":471,"height":160}]}
 {"event":"source_updated","timestamp":"2026-04-10T14:03:31.515Z","sources":[{"export":"PrimaryButtons","position":{"x":1046,"y":1831}},{"export":"DangerButtons","position":{"x":417,"y":1058},"width":471,"height":160}]}
+{"event":"widget_added","timestamp":"2026-04-10T14:21:22.880Z","widget":{"id":"image-sy9dfb","type":"image","position":{"x":1913,"y":2668},"props":{"src":"button-patterns--2026-04-10--16-21-22.png","private":false,"width":461,"height":154}}}
+{"event":"widget_removed","timestamp":"2026-04-10T14:21:49.293Z","widgetId":"image-sy9dfb"}
